Separate tools from virtual file system
- separates the built-in "ls" and "view" tools
- allows the tools to be brought in independently
- sets up a pattern for developers to build their own tools
