+++ assistants/codespace-assistant/assistant/chat.py
@@ -181,7 +181,6 @@ async def on_message_created(
         try:
             await respond_to_conversation(
                 message=message,
-                attachments_extension=attachments_extension,
                 context=context,
                 config=config,
                 metadata=metadata,
