+++ libraries/python/assistant-evaluations/.vscode/settings.json
@@ -19,7 +19,7 @@
   "python.analysis.inlayHints.functionReturnTypes": true,
   "python.analysis.typeCheckingMode": "standard",
   "python.defaultInterpreterPath": "${workspaceFolder}/.venv",
-  "python.testing.pytestEnabled": true,
+  "python.testing.pytestEnabled": false,
   "[python]": {
     "editor.defaultFormatter": "charliermarsh.ruff",
     "editor.formatOnSave": true,
