+++ libraries/python/chat-context-toolkit/test/virtual_filesystem/test_virtual_filesystem.py
@@ -9,11 +9,6 @@ from chat_context_toolkit.virtual_filesystem import (
     FileEntry,
     MountPoint,
     VirtualFileSystem,
-    WriteToolDefinition,
-)
-from openai.types.chat import (
-    ChatCompletionMessageToolCallParam,
-    ChatCompletionToolParam,
 )
 
 
@@ -24,11 +19,6 @@ class MockFileSource:
         """Initialize with files and directories."""
         self.files = files or {}
         self.directories = directories or {}
-
-    @property
-    def write_tools(self) -> Iterable[WriteToolDefinition]:
-        """Get write tools."""
-        return []
 
     async def list_directory(self, path: str) -> Iterable[DirectoryEntry | FileEntry]:
         """List directory contents."""
@@ -74,14 +64,15 @@ async def test_list_with_no_mounts():
 
 async def test_mount_and_list_root_directory():
     """Test mounting file source and listing root shows mounted directories."""
-    vfs = VirtualFileSystem()
     source = MockFileSource(directories={"/": ["file1.txt"]})
 
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=source,
-        )
+    vfs = VirtualFileSystem(
+        mounts=[
+            MountPoint(
+                entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
+                file_source=source,
+            )
+        ]
     )
 
     # Root directory should show mounted paths
@@ -95,19 +86,20 @@ async def test_mount_and_list_root_directory():
 
 async def test_duplicate_mount_raises_error():
     """Test mounting the same source multiple times."""
-    vfs = VirtualFileSystem()
     source = MockFileSource(directories={"/": ["file1.txt"]})
 
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=source,
-        )
+    vfs = VirtualFileSystem(
+        mounts=[
+            MountPoint(
+                entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
+                file_source=source,
+            )
+        ]
     )
 
     # Attempting to mount the same source again should raise an error
     with pytest.raises(ValueError, match="already mounted"):
-        vfs.mount(
+        vfs._mount(
             MountPoint(
                 entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
                 file_source=source,
@@ -117,47 +109,56 @@ async def test_duplicate_mount_raises_error():
 
 async def test_invalid_mount_paths_raise_error():
     """Test that mounting with invalid paths raises ValueError."""
-    vfs = VirtualFileSystem()
 
     # Attempt to mount with invalid paths
     with pytest.raises(ValueError, match="is invalid"):
-        vfs.mount(
-            MountPoint(
-                entry=DirectoryEntry(path="", description="Test", permission="read"),
-                file_source=MockFileSource(),
-            )
+        VirtualFileSystem(
+            mounts=[
+                MountPoint(
+                    entry=DirectoryEntry(path="", description="Test", permission="read"),
+                    file_source=MockFileSource(),
+                )
+            ]
         )
 
     with pytest.raises(ValueError, match="is invalid"):
-        vfs.mount(
-            MountPoint(
-                entry=DirectoryEntry(path="/", description="Test", permission="read"),
-                file_source=MockFileSource(),
-            )
+        VirtualFileSystem(
+            mounts=[
+                MountPoint(
+                    entry=DirectoryEntry(path="/", description="Test", permission="read"),
+                    file_source=MockFileSource(),
+                )
+            ]
         )
 
     with pytest.raises(ValueError, match="is invalid"):
-        vfs.mount(
-            MountPoint(
-                entry=DirectoryEntry(path="docs", description="Test", permission="read"),
-                file_source=MockFileSource(),
-            )
+        VirtualFileSystem(
+            mounts=[
+                MountPoint(
+                    entry=DirectoryEntry(path="docs", description="Test", permission="read"),
+                    file_source=MockFileSource(),
+                )
+            ]
         )
 
     with pytest.raises(ValueError, match="is invalid"):
-        vfs.mount(
-            MountPoint(
-                entry=DirectoryEntry(path="/docs/sub-dir", description="Test", permission="read"),
-                file_source=MockFileSource(),
-            )
+        VirtualFileSystem(
+            mounts=[
+                MountPoint(
+                    entry=DirectoryEntry(path="/docs/sub-dir", description="Test", permission="read"),
+                    file_source=MockFileSource(),
+                )
+            ]
         )
 
     with pytest.raises(ValueError, match="is invalid"):
-        vfs.mount(
-            MountPoint(
-                entry=DirectoryEntry(path="\\docs", description="Test", permission="read"),
-                file_source=MockFileSource(),
-            )
+        VirtualFileSystem(
+            mounts=[
+                MountPoint(
+                    entry=DirectoryEntry(path="\\docs", description="Test", permission="read"),
+                    file_source=MockFileSource(),
+                )
+            ]
         )
 
 
@@ -166,7 +167,7 @@ async def test_mount_and_list_mounted_directory_with_files():
     vfs = VirtualFileSystem()
     source = MockFileSource(files={"/file1.txt": "content1"}, directories={"/": ["file1.txt"]})
 
-    vfs.mount(
+    vfs._mount(
         MountPoint(
             entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
             file_source=source,
@@ -190,7 +191,7 @@ async def test_mount_and_list_mounted_directory_with_sub_directories():
         directories={"/": ["file1.txt", "subdir/"], "/subdir": ["file2.txt"]},
     )
 
-    vfs.mount(
+    vfs._mount(
         MountPoint(
             entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
             file_source=source,
@@ -218,7 +219,7 @@ async def test_mount_and_list_sub_directory():
         directories={"/": ["file1.txt", "subdir/"], "/subdir": ["file2.txt"]},
     )
 
-    vfs.mount(
+    vfs._mount(
         MountPoint(
             entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
             file_source=source,
@@ -252,7 +253,7 @@ async def test_read_file_from_mounted_source():
     vfs = VirtualFileSystem()
     source = MockFileSource(files={"/file1.txt": "hello world"})
 
-    vfs.mount(
+    vfs._mount(
         MountPoint(
             entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
             file_source=source,
@@ -273,13 +274,13 @@ async def test_multiple_mounts():
     docs_source = MockFileSource(files={"/readme.txt": "docs content"}, directories={"/": ["readme.txt"]})
     code_source = MockFileSource(files={"/main.py": "print('hello')"}, directories={"/": ["main.py"]})
 
-    vfs.mount(
+    vfs._mount(
         MountPoint(
             entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
             file_source=docs_source,
         )
     )
-    vfs.mount(
+    vfs._mount(
         MountPoint(
             entry=DirectoryEntry(path="/src", description="Test source code", permission="read"),
             file_source=code_source,
@@ -309,29 +310,11 @@ async def test_multiple_mounts():
     assert src_result[0].path == "/src/main.py"
 
 
-async def test_unmount_removes_access():
-    """Test unmounting removes access to file source."""
-    vfs = VirtualFileSystem()
-    source = MockFileSource(files={"/file1.txt": "content"}, directories={"/": ["file1.txt"]})
-
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=source,
-        )
-    )
-    vfs.unmount("/docs")
-
-    # After unmounting, directory should not be accessible
-    with pytest.raises(FileNotFoundError):
-        await vfs.list_directory("/docs")
-
-
 async def test_read_file_not_found_errors():
     """Test proper FileNotFoundError for non-existent paths."""
     vfs = VirtualFileSystem()
     source = MockFileSource(files={"/file1.txt": "content"}, directories={"/": ["file1.txt"]})
-    vfs.mount(
+    vfs._mount(
         MountPoint(
             entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
             file_source=source,
@@ -349,188 +332,6 @@ async def test_read_file_not_found_errors():
     # Non-existent file in mounted source
     with pytest.raises(FileNotFoundError):
         await vfs.read_file("/docs/nonexistent.txt")
-
-
-def test_tools_integration_with_mounted_sources():
-    """Test that tools from mounted sources are available."""
-    vfs = VirtualFileSystem()
-
-    # Create a file source with write tools
-    class SourceWithTools:
-        @property
-        def write_tools(self):
-            return [MockWriteTool()]
-
-        async def list_directory(self, path):
-            return []
-
-        async def read_file(self, path):
-            return ""
-
-    class MockWriteTool:
-        @property
-        def tool_param(self) -> ChatCompletionToolParam:
-            return {
-                "type": "function",
-                "function": {"name": "write_file", "description": "Write content to a file"},
-            }
-
-        async def execute(self, args: dict) -> str:
-            return "File written"
-
-    # Initially no tools
-    assert "write_file" not in vfs.tools
-
-    # After mounting source with tools, they should be available
-    source = SourceWithTools()
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=source,
-        )
-    )
-
-    # Tools should now include the mounted source's tools
-    assert "write_file" in vfs.tools
-
-
-async def test_execute_nonexistent_tool_raises_error():
-    """Test that executing a non-existent tool raises ValueError."""
-    vfs = VirtualFileSystem()
-
-    tool_call = ChatCompletionMessageToolCallParam(
-        id="test", function={"name": "nonexistent_tool", "arguments": "{}"}, type="function"
-    )
-
-    # Non-existent tool should raise ValueError
-    with pytest.raises(ValueError, match="Tool not found"):
-        await vfs.execute_tool(tool_call)
-
-
-async def test_execute_tool_delegates_to_source():
-    """Test that tool execution properly delegates to the appropriate mounted source."""
-    vfs = VirtualFileSystem()
-
-    # Create a file source with a write tool
-    class SourceWithTool:
-        @property
-        def write_tools(self):
-            return [TestWriteTool()]
-
-        async def list_directory(self, path):
-            return []
-
-        async def read_file(self, path):
-            return ""
-
-    class TestWriteTool:
-        @property
-        def tool_param(self) -> ChatCompletionToolParam:
-            return {
-                "type": "function",
-                "function": {"name": "test_write", "description": "Test write tool"},
-            }
-
-        async def execute(self, args: dict) -> str:
-            return f"Tool executed with args: {args}"
-
-    # Mount source with tool
-    source = SourceWithTool()
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=source,
-        )
-    )
-
-    # Execute the tool
-    tool_call = ChatCompletionMessageToolCallParam(
-        id="test",
-        function={"name": "test_write", "arguments": '{"content": "hello", "filename": "test.txt"}'},
-        type="function",
-    )
-
-    result = await vfs.execute_tool(tool_call)
-
-    # Should delegate to source and return result
-    assert result == "Tool executed with args: {'content': 'hello', 'filename': 'test.txt'}"
-
-
-async def test_multiple_sources_with_tools():
-    """Test that tools from multiple mounted sources are all available."""
-    vfs = VirtualFileSystem()
-
-    # Create two sources with different tools
-    class SourceWithTool1:
-        @property
-        def write_tools(self):
-            return [Tool1()]
-
-        async def list_directory(self, path):
-            return []
-
-        async def read_file(self, path):
-            return ""
-
-    class SourceWithTool2:
-        @property
-        def write_tools(self):
-            return [Tool2()]
-
-        async def list_directory(self, path):
-            return []
-
-        async def read_file(self, path):
-            return ""
-
-    class Tool1:
-        @property
-        def tool_param(self) -> ChatCompletionToolParam:
-            return {"type": "function", "function": {"name": "tool1", "description": "Tool 1"}}
-
-        async def execute(self, args: dict) -> str:
-            return "tool1 executed"
-
-    class Tool2:
-        @property
-        def tool_param(self) -> ChatCompletionToolParam:
-            return {"type": "function", "function": {"name": "tool2", "description": "Tool 2"}}
-
-        async def execute(self, args: dict) -> str:
-            return "tool2 executed"
-
-    starting_tools_count = len(vfs.tools)
-
-    # Mount both sources
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=SourceWithTool1(),
-        )
-    )
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/code", description="Test code", permission="read"),
-            file_source=SourceWithTool2(),
-        )
-    )
-
-    # Both tools should be available
-    tools = vfs.tools
-    assert len(tools) == starting_tools_count + 2
-    assert "tool1" in tools
-    assert "tool2" in tools
-
-    # Both tools should be executable
-    result1 = await vfs.execute_tool(
-        ChatCompletionMessageToolCallParam(id="test1", function={"name": "tool1", "arguments": "{}"}, type="function")
-    )
-    assert result1 == "tool1 executed"
-
-    result2 = await vfs.execute_tool(
-        ChatCompletionMessageToolCallParam(id="test2", function={"name": "tool2", "arguments": "{}"}, type="function")
-    )
-    assert result2 == "tool2 executed"
 
 
 async def test_source_error_propagation():
@@ -552,7 +353,7 @@ async def test_source_error_propagation():
                 raise FileNotFoundError("File not found in source")
             return "content"
 
-    vfs.mount(
+    vfs._mount(
         MountPoint(
             entry=DirectoryEntry(path="/test", description="Test error source", permission="read"),
             file_source=ErrorSource(),
@@ -565,242 +366,3 @@ async def test_source_error_propagation():
 
     with pytest.raises(FileNotFoundError, match="File not found in source"):
         await vfs.read_file("/test/error.txt")
-
-
-async def test_unmount_removes_tools():
-    """Test that unmounting a source removes its tools."""
-    vfs = VirtualFileSystem()
-
-    class SourceWithTool:
-        @property
-        def write_tools(self):
-            return [TestTool()]
-
-        async def list_directory(self, path):
-            return []
-
-        async def read_file(self, path):
-            return ""
-
-    class TestTool:
-        @property
-        def tool_param(self) -> ChatCompletionToolParam:
-            return {"type": "function", "function": {"name": "test_tool", "description": "Test"}}
-
-        async def execute(self, args: dict) -> str:
-            return "executed"
-
-    starting_tools_count = len(vfs.tools)
-
-    # Mount source with tool
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=SourceWithTool(),
-        )
-    )
-    assert len(vfs.tools) == starting_tools_count + 1
-    assert "test_tool" in vfs.tools
-
-    # Unmount should remove tools
-    vfs.unmount("/docs")
-    assert len(vfs.tools) == starting_tools_count
-
-    # Tool should no longer be executable
-    with pytest.raises(ValueError, match="Tool not found"):
-        await vfs.execute_tool(
-            ChatCompletionMessageToolCallParam(
-                id="test", function={"name": "test_tool", "arguments": "{}"}, type="function"
-            )
-        )
-
-
-def test_builtin_ls_and_view_tools_available():
-    """Test that VFS provides built-in ls and view tools."""
-    vfs = VirtualFileSystem()
-
-    # Built-in tools should be available even with no mounts
-    tools = vfs.tools
-    assert "ls" in tools
-    assert "view" in tools
-
-    # Tools should have proper definitions
-    ls_tool = tools["ls"]
-    assert ls_tool["type"] == "function"
-    assert ls_tool["function"]["name"] == "ls"
-
-    view_tool = tools["view"]
-    assert view_tool["type"] == "function"
-    assert view_tool["function"]["name"] == "view"
-
-
-async def test_builtin_ls_tool_lists_directory():
-    """Test that the built-in ls tool can list directory contents."""
-    vfs = VirtualFileSystem()
-
-    # Create a source with mixed permissions and file types
-    class TestFileSource:
-        async def list_directory(self, path: str):
-            if path == "/":
-                return [
-                    FileEntry(
-                        path="/writable-file.txt",
-                        size=100,
-                        timestamp=datetime.now(),
-                        permission="read_write",
-                        description="A writable file",
-                    ),
-                    FileEntry(
-                        path="/readonly-file.txt",
-                        size=50,
-                        timestamp=datetime.now(),
-                        permission="read",
-                        description="A read-only file",
-                    ),
-                    DirectoryEntry(path="/sub-dir-1", description="A sub-directory", permission="read"),
-                ]
-            elif path == "/sub-dir-1":
-                return [
-                    FileEntry(
-                        path="/sub-dir-1/nested-file.txt",
-                        size=25,
-                        timestamp=datetime.now(),
-                        permission="read",
-                        description="A nested file",
-                    ),
-                ]
-            else:
-                raise FileNotFoundError(f"Directory not found: {path}")
-
-        async def read_file(self, path: str) -> str:
-            return "file content"
-
-        @property
-        def write_tools(self):
-            return []
-
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=TestFileSource(),
-        )
-    )
-
-    # Test ls on root (should show mounted directories)
-    ls_call = ChatCompletionMessageToolCallParam(
-        id="test", function={"name": "ls", "arguments": '{"path": "/"}'}, type="function"
-    )
-    result = await vfs.execute_tool(ls_call)
-    assert isinstance(result, str)
-    expected_root = "\n".join(["List of files in /:", "dr--    - docs/ - Test docs"])
-    assert result == expected_root
-
-    # Test ls on mounted directory
-    ls_call = ChatCompletionMessageToolCallParam(
-        id="test", function={"name": "ls", "arguments": '{"path": "/docs"}'}, type="function"
-    )
-    result = await vfs.execute_tool(ls_call)
-    assert isinstance(result, str)
-    expected = "\n".join([
-        "List of files in /docs:",
-        "dr--    - sub-dir-1/ - A sub-directory",
-        "-r--  50B readonly-file.txt - A read-only file",
-        "-rw- 100B writable-file.txt - A writable file",
-    ])
-    assert result == expected
-
-    # Test ls on subdirectory
-    ls_call = ChatCompletionMessageToolCallParam(
-        id="test", function={"name": "ls", "arguments": '{"path": "/docs/sub-dir-1"}'}, type="function"
-    )
-    result = await vfs.execute_tool(ls_call)
-    assert isinstance(result, str)
-    expected_nested = "\n".join(["List of files in /docs/sub-dir-1:", "-r--  25B nested-file.txt - A nested file"])
-    assert result == expected_nested
-
-
-async def test_builtin_view_tool_reads_file():
-    """Test that the built-in view tool can read file contents."""
-    vfs = VirtualFileSystem()
-    source = MockFileSource(files={"/file1.txt": "Hello World", "/subdir/file2.txt": "Nested Content"})
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=source,
-        )
-    )
-
-    # Test view on file
-    view_call = ChatCompletionMessageToolCallParam(
-        id="test", function={"name": "view", "arguments": '{"path": "/docs/file1.txt"}'}, type="function"
-    )
-    result = await vfs.execute_tool(view_call)
-    assert result == "<file path=/docs/file1.txt>\nHello World\n</file>"
-
-    # Test view on nested file
-    view_call = ChatCompletionMessageToolCallParam(
-        id="test", function={"name": "view", "arguments": '{"path": "/docs/subdir/file2.txt"}'}, type="function"
-    )
-    result = await vfs.execute_tool(view_call)
-    assert result == "<file path=/docs/subdir/file2.txt>\nNested Content\n</file>"
-
-
-async def test_builtin_tools_error_handling():
-    """Test error handling for built-in ls and view tools."""
-    vfs = VirtualFileSystem()
-
-    # ls on non-existent path should raise error
-    ls_call = ChatCompletionMessageToolCallParam(
-        id="test", function={"name": "ls", "arguments": '{"path": "/nonexistent"}'}, type="function"
-    )
-    result = await vfs.execute_tool(ls_call)
-    assert result == "Error: Directory not found: /nonexistent"
-
-    # view on non-existent file should raise error
-    view_call = ChatCompletionMessageToolCallParam(
-        id="test", function={"name": "view", "arguments": '{"path": "/nonexistent.txt"}'}, type="function"
-    )
-    result = await vfs.execute_tool(view_call)
-    assert (
-        result
-        == "Error: File at path /nonexistent.txt not found. Please pay attention to the available files and try again."
-    )
-
-
-def test_builtin_tools_combined_with_source_tools():
-    """Test that built-in tools work alongside tools from mounted sources."""
-    vfs = VirtualFileSystem()
-
-    # Create source with custom tool
-    class SourceWithTool:
-        @property
-        def write_tools(self):
-            return [CustomTool()]
-
-        async def list_directory(self, path):
-            return []
-
-        async def read_file(self, path):
-            return ""
-
-    class CustomTool:
-        @property
-        def tool_param(self) -> ChatCompletionToolParam:
-            return {"type": "function", "function": {"name": "custom_tool", "description": "Custom tool"}}
-
-        async def execute(self, args: dict) -> str:
-            return "custom executed"
-
-    vfs.mount(
-        MountPoint(
-            entry=DirectoryEntry(path="/docs", description="Test docs", permission="read"),
-            file_source=SourceWithTool(),
-        )
-    )
-
-    # Should have both built-in and source tools
-    tools = vfs.tools
-    assert "ls" in tools  # Built-in
-    assert "view" in tools  # Built-in
-    assert "custom_tool" in tools  # From source
-    assert len(tools) == 3
